cpu scheduler: make _reciprocal_shares_times_2_32 wider to avoid overflow problems
reciprocal shares calculation[1] results in overflow if shares is 1.0f, which
will later mess with calculation of vruntime.

[1]: _reciprocal_shares_times_2_power_32((uint64_t(1) << 32) / _shares)

Message-Id: <20171013175609.7021-1-raphaelsc@scylladb.com>

+++ core/reactor.hh
@@ -732,7 +732,7 @@ private:
         explicit task_queue(unsigned id, sstring name, float shares);
         int64_t _vruntime = 0;
         float _shares;
-        unsigned _reciprocal_shares_times_2_power_32;
+        int64_t _reciprocal_shares_times_2_power_32;
         bool _current = false;
         bool _active = false;
         uint8_t _id;
